+++ client/src/app/services/objetivo.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ServerService } from './server.service';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
@@ -9,39 +10,47 @@ export class ObjetivoService {
   headers = {headers: new HttpHeaders().set('Content-Type', 'application/json').set('Authorization', 'Bearer '+localStorage.getItem('token'))};
   api_url
 
-  constructor(private api:ServerService, private http:HttpClient) {
+  constructor(private api:ServerService, private http:HttpClient, private auth:AuthService) {
     this.api_url = api.API_URI + 'objetivo/'
   }
 
   crear(objetivo){
+    if(this.auth.isTokenValid())
     return this.http.post(this.api_url, objetivo, this.headers)
   }
 
   crearGrupo(objetivos){
-    return this.http.post(this.api_url+'bulk', objetivos, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.post(this.api_url+'bulk', objetivos, this.headers)
   }
 
   listar(body){
-    return this.http.post(this.api_url+'list', body, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.post(this.api_url+'list', body, this.headers)
   }
 
   ver(id){
-    return this.http.get(this.api_url+id, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.get(this.api_url+id, this.headers)
   }
 
   actualizar(id, objetivo){
-    return this.http.put(this.api_url+id, objetivo, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.put(this.api_url+id, objetivo, this.headers)
   }
 
   borrar(id){
-    return this.http.delete(this.api_url+id, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.delete(this.api_url+id, this.headers)
   }
 
   asociar(id, creencias){
-    return this.http.post(this.api_url+id+'/asociar', creencias, this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.post(this.api_url+id+'/asociar', creencias, this.headers)
   }
 
   verAsociado(id){
-    return this.http.get(this.api_url+id+'/asociado', this.headers)
+    if(this.auth.isTokenValid())
+      return this.http.get(this.api_url+id+'/asociado', this.headers)
   }
 }
